// FrontEnd/src/app/services/api-backend/model/bodega.ts
/**
 * Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Lote } from './lote';
import { Sucursale } from './sucursale';


export interface Bodega { 
    idBodega?: number;
    idSucursal?: number;
    nombre?: string | null;
    idSucursalNavigation?: Sucursale;
    lotes?: Array<Lote> | null;
}