// FrontEnd/src/app/services/api-backend/model/lote.ts
/**
 * Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { AplicacionesMedicamento } from './aplicacionesMedicamento';
import { Bodega } from './bodega';
import { Producto } from './producto';
import { ProductosFactura } from './productosFactura';


export interface Lote { 
    idLote?: number;
    idProducto?: number;
    idBodega?: number;
    exitencia?: number;
    estado?: string | null;
    precioUnitario?: number;
    marca?: string | null;
    fechaIngreso?: string;
    fechaCaducidad?: string;
    idBodegaNavigation?: Bodega;
    idProductoNavigation?: Producto;
    aplicacionesMedicamentos?: Array<AplicacionesMedicamento> | null;
    productosFacturas?: Array<ProductosFactura> | null;
}